**CRITICAL**: Fixed a bug where the check for already existing files was only performed for regular files, but not directories. Fixed a bug which would result in incorrect handling of files located in the root directory (`/`) if passed as absolute paths (`/1-file /2-file ...`). Duplicate forward slashes in paths are now removed as appropriate.
Improved the method of quoting strings for displaying them in messages.
All instances of using `\'"${parameter//\'/\'\\\'\'}"\'` were replaced
with `"${parameter@Q}"`.
Improved code for the
`` Use of capture groups in regular expressions set with `-b, --before` and `-a, --after` is not allowed ``
error message. Removed `sed` dependency.

Improved formatting of the built-in help message.
Added information about dependencies to README.
Added a changelog.

Removed a superfluous `IFS=' '` from the `getopt` line.
Removed a superfluous `:` on the line preceding the `mv` command.